"Add Element-Plus and auto-import plugins to project & update payment process functionality"
This commit changes the Vite configuration and the payment processing logic in the ButtonCard.vue and util.js components. It imports and configures the AutoImport and VueComponents plugins from unplugin to enable auto-importing of components and utilities from Element-Plus library.

In ButtonCard.vue, the payment's request body and error handling have been modified. For util.js, the function makePayment is updated to return the promise. This allows more flexibility in handling the response outside the function.

This changes were made to improve the code readability and flexibility, as well as performance in terms of bundle size by only importing what's needed.